refactor: split the tables into different files
